Made Navb solo component
