+++ src/App/Nav/Nav.css
@@ -1,0 +1,18 @@
+.nav-container {
+  width: 25%;
+  height: 100%;
+}
+
+.hamburger {
+  width: 100%;
+  height: 100%;
+  color: #AE785E;
+}
+
+.menu {
+  background-color: #B19271;
+}
+
+.menu-items {
+ color: #84695D;
+}
